+++ go/worker/BUILD.bazel
@@ -11,14 +11,10 @@ go_library(
     deps = [
         "//go/base/blobstore:go_default_library",
         "//go/base/blobstore/minio:go_default_library",
+        "//go/cauldron/worker/http:go_default_library",
+        "//go/cauldron/worker/activities/rayhttp:go_default_library",
+        "//go/cauldron/worker/activities/sparkhttp:go_default_library",
         "//go/worker/activities:go_default_library",
-        "//go/worker/activities/http:go_default_library",
-        "//go/worker/activities/rayhttp:go_default_library",
-        "//go/worker/plugins/rayhttp:go_default_library",
-        "//go/worker/ray:go_default_library",
-        "//go/worker/activities/sparkhttp:go_default_library",
-        "//go/worker/plugins/sparkhttp:go_default_library",
-        "//go/worker/spark:go_default_library",
         "//go/worker/starlark:go_default_library",
         "//go/worker/workflowfx:go_default_library",
         "//go/worker/workflows:go_default_library",
@@ -28,8 +24,6 @@ go_library(
         "@org_uber_go_config//:go_default_library",
         "@org_uber_go_fx//:go_default_library",
         "@org_uber_go_yarpc//:go_default_library",
-        "@org_uber_go_yarpc//api/transport:go_default_library",
         "@org_uber_go_yarpc//transport/grpc:go_default_library",
-        "@org_uber_go_yarpc//transport/http:go_default_library",
     ],
 )
