+++ go/worker/activities/BUILD.bazel
@@ -6,11 +6,9 @@ go_library(
     importpath = "github.com/michelangelo-ai/michelangelo/go/worker/activities",
     visibility = ["//visibility:public"],
     deps = [
+        "//go/cauldron/worker/activities/rayhttp:go_default_library",
+        "//go/cauldron/worker/activities/sparkhttp:go_default_library",
         "//go/worker/activities/cachedoutput:go_default_library",
-        "//go/worker/activities/ray:go_default_library",
-        "//go/worker/activities/rayhttp:go_default_library",
-        "//go/worker/activities/spark:go_default_library",
-        "//go/worker/activities/sparkhttp:go_default_library",
         "//go/worker/activities/storage:go_default_library",
         "@org_uber_go_fx//:go_default_library",
     ],
