--- go/worker/activities/http/BUILD.bazel
@@ -1,13 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "config.go",
-    ],
-    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/activities/http",
-    visibility = ["//visibility:public"],
-    deps = [
-        "@org_uber_go_fx//:go_default_library",
-    ],
-)
