--- go/worker/activities/rayhttp/BUILD.bazel
@@ -1,25 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "module.go",
-        "ray_activities.go",
-    ],
-    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/activities/rayhttp",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//go/worker/activities/http:go_default_library",
-        "//go/worker/ray:go_default_library",
-        "@com_github_cadence_workflow_starlark_worker//activity:go_default_library",
-        "@com_github_cadence_workflow_starlark_worker//service:go_default_library",
-        "@com_github_cadence_workflow_starlark_worker//workflow:go_default_library",
-        "@com_github_cadence_workflow_starlark_worker//worker:go_default_library",
-        "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
-        "@io_k8s_apimachinery//pkg/apis/meta/v1/unstructured:go_default_library",
-        "@io_k8s_sigs_controller_runtime//pkg/log:go_default_library",
-        "@org_uber_go_fx//:go_default_library",
-        "@org_uber_go_yarpc//yarpcerrors:go_default_library",
-        "@org_uber_go_zap//:go_default_library",
-    ],
-)
