--- go/worker/plugins/rayhttp/plugin.go
@@ -1,25 +1,0 @@
-package rayhttp
-
-import (
-	"github.com/cadence-workflow/starlark-worker/service"
-	"github.com/cadence-workflow/starlark-worker/worker"
-	"go.starlark.net/starlark"
-)
-
-const pluginID = "rayhttp"
-
-var Plugin = &plugin{}
-
-type plugin struct{}
-
-var _ service.IPlugin = (*plugin)(nil)
-
-func (r *plugin) ID() string {
-	return pluginID
-}
-
-func (r *plugin) Create(_ service.RunInfo) starlark.Value {
-	return newModule()
-}
-
-func (r *plugin) Register(_ worker.Registry) {}
