--- go/worker/plugins/sparkhttp/BUILD.bazel
@@ -1,22 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "plugin.go",
-        "starlark_module.go",
-    ],
-    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/plugins/sparkhttp",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//go/worker/activities/sparkhttp:go_default_library",
-        "//go/worker/plugins/utils:go_default_library",
-        "//go/worker/spark:go_default_library",
-        "@com_github_cadence_workflow_starlark_worker//ext:go_default_library",
-        "@com_github_cadence_workflow_starlark_worker//service:go_default_library",
-        "@com_github_cadence_workflow_starlark_worker//worker:go_default_library",
-        "@com_github_cadence_workflow_starlark_worker//workflow:go_default_library",
-        "@net_starlark_go//starlark:go_default_library",
-        "@org_uber_go_zap//:go_default_library",
-    ],
-)
