--- go/worker/plugins/sparkhttp/plugin.go
@@ -1,25 +1,0 @@
-package sparkhttp
-
-import (
-	"github.com/cadence-workflow/starlark-worker/service"
-	"github.com/cadence-workflow/starlark-worker/worker"
-	"go.starlark.net/starlark"
-)
-
-const pluginID = "sparkhttp"
-
-var Plugin = &plugin{}
-
-type plugin struct{}
-
-var _ service.IPlugin = (*plugin)(nil)
-
-func (r *plugin) ID() string {
-	return pluginID
-}
-
-func (r *plugin) Create(_ service.RunInfo) starlark.Value {
-	return newModule()
-}
-
-func (r *plugin) Register(_ worker.Registry) {}
