+++ go/worker/starlark/BUILD.bazel
@@ -6,11 +6,9 @@ go_library(
     importpath = "github.com/michelangelo-ai/michelangelo/go/worker/starlark",
     visibility = ["//visibility:public"],
     deps = [
+        "//go/cauldron/worker/plugins/rayhttp:go_default_library",
+        "//go/cauldron/worker/plugins/sparkhttp:go_default_library",
         "//go/worker/plugins/cachedoutput:go_default_library",
-        "//go/worker/plugins/ray:go_default_library",
-        "//go/worker/plugins/rayhttp:go_default_library",
-        "//go/worker/plugins/spark:go_default_library",
-        "//go/worker/plugins/sparkhttp:go_default_library",
         "//go/worker/plugins/storage:go_default_library",
         "@com_github_cadence_workflow_starlark_worker//service:go_default_library",
         "@com_github_cadence_workflow_starlark_worker//worker:go_default_library",
